wrn--
SVN revision: 59197

+++ legacy/elementary/src/lib/elm_segment_control.c
@@ -235,7 +235,7 @@ _segment_on(Elm_Segment_Item *it)
    edje_object_signal_emit(it->base.view, "elm,state,segment,selected", "elm");
 
    wd->selected_item = it;
-   evas_object_smart_callback_call(wd->obj, SIG_CHANGED, (void*) it->seg_index);
+   evas_object_smart_callback_call(wd->obj, SIG_CHANGED, (void *)(unsigned long)it->seg_index);
 }
 
 static void
